added set subpixel
Added method to set a sub pixel
Added an example that handles 6 strings of 2 leds, updating only
specific subpixels and syncs the string only if needed

I found some code online that promises an accurate clock, so decided I
wanted to try it out.
Meshing it with RGB pixels
The code still needs the ability to be set manually to a specific time.
A simple hold analog button press to forward the minutes should do
(speeding up the minutes if long pressed).